ishy-bot/requirements.txt
Flask==2.3.4
APScheduler==3.10.1
requests==2.31.0
openai==1.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
PyGithub==1.78
